+++ CHANGES.md
@@ -7,6 +7,7 @@ dev
 Improvements:
 - Add default viewpoint (#61)
 - Improve performance of `uniformPolyhedron`s' edges (#62)
+- Improve performance of coplanar polygons (#63)
 
 Documentation:
 - Add more examples (#59)